Move stateless components into container directory
