+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Homepage from './components/homepage/Homepage';
-import CountryD from './components/detailsPage/CountryD';
+import CountryD from './components/container/detailsPage/CountryD';
 import './App.css';
 
 function App() {
