+++ src/components/container/detailsPage/CountryD.css
