+++ src/components/homepage/Homepage.js
@@ -2,7 +2,7 @@ import Container from '@mui/material/Container';
 import Countries from './Country';
 import './Homepage.css';
 import Banner from './Banner';
-import Navbar from '../navbar/Navbar';
+import Navbar from '../container/navbar/Navbar';
 
 const Homepage = () => (
   <Container className="main-container">
